html/dom/elements/global-attributes/the-translate-attribute/the-translate-attribute-012.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>translate=""</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://html.spec.whatwg.org/multipage/dom.html#the-translate-attribute'>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" type="text/css" href="../../../../../testharness.css">
<meta name="assert" content="If the translate attribute is set to a null string, javascript will detect the translation mode of text as translate-enabled.">
</head>
<body>
<div class="test"><div id="box" translate="">&nbsp;</div></div>
<script>
test(function() {
assert_true(document.getElementById('box').translate);
}, " ");
</script>
<div id='log'></div>
</body>
</html>
